Backend/tansacs/requirement.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
boto3==1.29.6
botocore==1.32.6
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
cssselect2==0.7.0
Django==4.2.7
django-cors-headers==4.3.1
django-storages==1.14.2
djangorestframework==3.14.0
fonttools==4.46.0
freetype-py==2.3.0
html5lib==1.1
idna==3.6
jmespath==1.0.1
lxml==4.9.3
oscrypto==1.3.0
Pillow==10.1.0
psycopg2==2.9.9
pycairo==1.25.1
pycparser==2.21
pydyf==0.8.0
pyHanko==0.21.0
pyhanko-certvalidator==0.26.2
pypdf==3.17.1
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.7
requests==2.31.0
rlPyCairo==0.3.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.2
urllib3==2.0.7
webencodings==0.5.1
xhtml2pdf==0.2.13
zopfli==0.2.3
